requirements.txt
fastapi
uvicorn
motor
pydantic[dotenv]
sentry-sdk
python-dotenv
